Adds dependecy repo and bootstrapping for sql files
Adds the ability to identify a dependencies repository. This repository
is assumed to have a /themes, /plugins, and /dropins directory in it
similar to how the /src folder is set up. However, this folder can be
placed into VCS to control project dependencies such as commercial
plugins and themes. Just like /src, these will be symlinked when the
setup script is triggered.

Also in this commit is some additional handling for sql file. During
setup, you can specify a local fild path with --db=<path>. In addition
when bootstrapping, the script will look for sql files in the current
directory and automatically move the first one found into /config/data
for import. Added support for sql.gz files, as well as simplified the
find and replace logic for multisite.